move-pool-configurator: Move pool_configurator into new model_runner_components/ subdir
Refactor chain ID: move-pool-configurator
