+++ test/registered/unit/model_executor/test_pool_configurator.py
@@ -28,7 +28,7 @@ def mock_cpu_env(kv_size=2, tp_size=1, swa_eviction_interval=4):
     with (
         patch("torch._utils._element_size", return_value=kv_size),
         patch(
-            "sglang.srt.model_executor.pool_configurator.get_attention_tp_size",
+            "sglang.srt.model_executor.model_runner_components.pool_configurator.get_attention_tp_size",
             return_value=tp_size,
         ),
         envs.SGLANG_SWA_EVICTION_INTERVAL.override(swa_eviction_interval),
@@ -163,7 +163,7 @@ class TestDefaultConfigurator(unittest.TestCase):
     def _run(self, available_bytes, page_size=1, **kwargs):
         mr = _make_model_runner(page_size=page_size, **kwargs)
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -219,7 +219,7 @@ class TestHybridSWAConfigurator(unittest.TestCase):
     def _run(self, available_bytes, **kwargs):
         mr = self._make_swa_runner(**kwargs)
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -313,7 +313,7 @@ class TestHybridSWAConfigurator(unittest.TestCase):
             disable_overlap_schedule=True,  # spec-v1: no double allocation
         )
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -352,7 +352,7 @@ class TestHybridSWAConfigurator(unittest.TestCase):
             disable_overlap_schedule=False,  # spec-v2: 2 * get_alloc_len_per_decode
         )
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -378,7 +378,7 @@ class TestHybridSWAConfigurator(unittest.TestCase):
             disaggregation_mode="decode",
         )
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -409,7 +409,7 @@ class TestHybridSWAConfigurator(unittest.TestCase):
             disable_overlap_schedule=False,  # overlap -> 2 chunks in flight
         )
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -441,7 +441,7 @@ class TestHybridSWAConfigurator(unittest.TestCase):
             disaggregation_decode_extra_slots=2,
         )
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -469,7 +469,7 @@ class TestAllSWAConfigurator(unittest.TestCase):
             **kwargs,
         )
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -519,7 +519,7 @@ class TestEagleConfigurator(unittest.TestCase):
         mr.eagle_draft_num_layers = eagle_draft_num_layers
 
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 create_memory_pool_configurator,
             )
 
@@ -536,7 +536,7 @@ class TestFactory(unittest.TestCase):
     def test_default_for_non_swa(self):
         mr = _make_model_runner(is_hybrid_swa=False)
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 DefaultPoolConfigurator,
                 create_memory_pool_configurator,
             )
@@ -552,7 +552,7 @@ class TestFactory(unittest.TestCase):
             swa_num_kv_heads=4,
         )
         with mock_cpu_env():
-            from sglang.srt.model_executor.pool_configurator import (
+            from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                 HybridSWAPoolConfigurator,
                 create_memory_pool_configurator,
             )
@@ -574,13 +574,13 @@ class TestFactory(unittest.TestCase):
                 max_running_requests=max_running_requests,
             )
             with mock_cpu_env():
-                from sglang.srt.model_executor.pool_configurator import (
+                from sglang.srt.model_executor.model_runner_components.pool_configurator import (
                     create_memory_pool_configurator,
                 )
 
                 return create_memory_pool_configurator(mr)
 
-        from sglang.srt.model_executor.pool_configurator import (
+        from sglang.srt.model_executor.model_runner_components.pool_configurator import (
             SWAChunkCapPoolConfigurator,
         )
